Topic Monitoring Improvements.
Can set the topic rate in the config. More convenient than throttling topics in launch files.
If the topic rate is not set, then sentor subscribes to the original topic (as it does normally).
Service names (for the `call` process) and topic names (for the `publish` process) can be retrieved from rosparams and environment variables. Sentor automatically checks the names provided in the config.
Processes are now not verbose by default.
Some minor improvements.