Custom Lambda Expressions.
You can now create custom lambda expressions in python files instead of specifying them explicitly in the config.
See the first lambda expression in the config `example.yaml`, which points to a function `CustomLambda` from file `CustomLambda.py` located in package `sentor`.
These are to be used when a lambda expression is too complex to be specified in the config explicitly.