Improvements.
For the `reconf` process you can now set the value of a param in the config to `_default` which reconfigures the param back to its original state.
The argument for the param's namespace has been changed from `ns` to `namespace`.
For topic mapping you no longer set the map and base frames as a rosparam (i.e in a launch file) but in the config instead, allowing maps using different tf transforms to be created simultaneously.
The topic map msg now includes a field for the base frame (`child_frame_id`).
For topic mapping the `stat` arg for the standard deviation has been changed from `std` to `stdev`.
Some general tidying.